Move comment processing to a separate file
