+++ novelwriter/core/index.py
@@ -43,6 +43,7 @@ from novelwriter.common import (
 from novelwriter.constants import nwFiles, nwKeyWords, nwStyles
 from novelwriter.enum import nwComment, nwItemClass, nwItemLayout, nwItemType
 from novelwriter.error import logException
+from novelwriter.text.comments import processComment
 from novelwriter.text.counting import standardCounter
 
 if TYPE_CHECKING:  # pragma: no cover
@@ -1343,48 +1344,3 @@ class IndexHeading:
                 else:
                     raise ValueError("The itemIndex contains an invalid reference type")
         return
-
-
-# Text Processing Functions
-# =========================
-
-MODIFIERS = {
-    "synopsis": nwComment.SYNOPSIS,
-    "short":    nwComment.SHORT,
-    "note":     nwComment.NOTE,
-    "footnote": nwComment.FOOTNOTE,
-}
-KEY_REQ = {
-    "synopsis": 0,  # Key not allowed
-    "short":    0,  # Key not allowed
-    "note":     1,  # Key optional
-    "footnote": 2,  # Key required
-}
-
-
-def _checkModKey(modifier: str, key: str) -> bool:
-    """Check if a modifier and key set are ok."""
-    if modifier in MODIFIERS:
-        if key == "":
-            return KEY_REQ[modifier] < 2
-        elif key.replace("_", "").isalnum():
-            return KEY_REQ[modifier] > 0
-    return False
-
-
-def processComment(text: str) -> tuple[nwComment, str, str, int, int]:
-    """Extract comment style, key and text. Should only be called on
-    text starting with a %.
-    """
-    if text[:2] == "%~":
-        return nwComment.IGNORE, "", text[2:].lstrip(), 0, 0
-
-    check = text[1:].strip()
-    start, _, content = check.partition(":")
-    modifier, _, key = start.rstrip().partition(".")
-    if content and (clean := modifier.lower()) and _checkModKey(clean, key):
-        col = text.find(":") + 1
-        dot = text.find(".", 0, col) + 1
-        return MODIFIERS[clean], key, content.lstrip(), dot, col
-
-    return nwComment.PLAIN, "", check, 0, 0
